summle,afl: allow to initialize with fuzzer
This extension allows to initialize the relevant arrays
that represent the positions to place the numbers.

This way, a fuzzer can dump data via stdin that is then
interpreted as the numbers to pick.

Note: this initialization only takes place in case CBMC is not used to
verify the program.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>